requirements.txt
python-telegram-bot==20.8
tweepy==4.14.0
openai==0.28
pydantic==1.10.13
python-dotenv
aiohttp
certifi
nest_asyncio
